+++ lib/app/modules/home/views/details/detail.controller.dart
@@ -81,8 +81,10 @@ class DetailController extends GetxController {
   }
 
   updateCartItems() {
-    cartItems$.value =
-        cartProducts$.map((e) => e.cartValue).reduce((a, b) => a + b);
+    cartItems$.value = cartProducts$
+        .map((e) => e.cartValue)
+        .reduce((a, b) => a + b.toInt())
+        .toInt();
     cartItems$.refresh();
   }
